refactor: split cli.ts into command modules
cli.ts was 2336 lines (82KB). Split into 12 focused modules
under src/commands/ grouped by help menu sections:

- core.ts: init, scan, status, verify, doctor
- transcript.ts: clean, watch, scan-history, clean-history
- secrets.ts: secret, run, env, import, setup
- rules.ts: custom deny rules
- git.ts: hook, scan-staged
- mcp.ts: protect-mcp, mcp-status, mcp-unprotect
- backend.ts: backend, migrate, cache
- broker.ts: credential broker daemon
- scope.ts: scope discovery
- session.ts: warm, install
- help.ts: help text
- utils.ts: shared formatters and VERSION

cli.ts is now a 142-line thin dispatcher. All 792 tests pass.
